requirements.txt
langchain
python-dotenv
ipykernel
langchain-community
pypdf
beautifulsoup4
arxiv
langchain-core
scikit-learn
wikipedia
langchain-text-splitters
chromadb
langchain-huggingface
sentence-transformers
faiss-cpu
langchain-chroma
streamlit
langchain-groq
uvicorn
fastapi
langserve
sse_starlette
arxiv
wikipedia
validators
youtube-transcript-api
pytube
unstructured
numexpr